#190 Provide workaround (read: it doesn't crash anymore; it's still gross) for invalid sub-model head nodes.

+++ src/client/renderer/r_types.h
@@ -152,7 +152,7 @@ typedef struct {
 	vec3_t mins, maxs;
 	vec3_t origin; // for sounds or lights
 	float radius;
-	int head_node;
+	unsigned short head_node;
 	unsigned short first_face, num_faces;
 } r_bsp_submodel_t;
